# source/user_service/openapi.yaml
openapi: 3.0.0
info:
  title: User service API
  version: 1.0.0
  description: Social network user service API

paths:
  /register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              required:
                - username
                - password
      responses:
        '201':
          description: Successful registration
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  username:
                    type: string
        '409':
          description: Username already exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
      
  /update:
    put:
      summary: Update user data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
                first_name:
                  type: string
                last_name:
                  type: string
                birth_day:
                  type: string
                  format: date
                email:
                  type: string
                phone_number:
                  type: string
              required:
                - username
                - password
      responses:
        '200':
          description: Successful update
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '404':
          description: Invalid username
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '401':
          description: Invalid password
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '500':
          description: Server internal exception
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string

  /users:
    get:
      summary: Get all users (for debug purposes)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                secret_key:
                  type: string
              required:
                - secret_key
      responses:
        '200':
          description: Get list of all users
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    username:
                      type: string
                    first_name:
                      type: string
                    last_name:
                      type: string
                    birth_day:
                      type: string
                      format: date
                    email:
                      type: string
                    phone_number:
                      type: string

  /login:
    post:
      summary: User authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              required:
                - username
                - password
      responses:
        '200':
          description: Successful authorization
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user_id:
                    type: integer
        '401':
          description: Invalid username or password
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string